move fn into utility file
